requirements/requirements.txt
falcon==1.0.0
gunicorn==19.6.0
gevent==1.1.1
python-dateutil>=0.0.0

pymongo==3.2.2

kafka-python>=0.0.0
requests==2.18
